Separate stream import for tracing
