+++ packages/common/codec-protobuf/src/proto/buffer-patch.ts
