+++ packages/common/codec-protobuf/src/proto/interface.ts
