+++ packages/common/codec-protobuf/src/proto/precompiled-mapping/codegen.ts
