+++ packages/common/codec-protobuf/src/proto/precompiled-mapping/create-message-mapper.ts
