+++ packages/common/codec-protobuf/src/proto/substitutions/any.ts
