+++ packages/common/codec-protobuf/src/proto/substitutions/struct.test.ts
