+++ packages/common/codec-protobuf/src/proto/substitutions/timestamp.ts
