+++ packages/common/codec-protobuf/src/stream/stream.test.ts
