+++ packages/common/codec-protobuf/src/stream/stream.ts
