+++ packages/common/codec-protobuf/src/stream/util.ts
